Basic saga refactoring
  - Reorganize sagas/utils imports
  - Move `doCheckTokenExpired` from sagas/login to sagas/utils
  - Sagas that need to query multiple times now use `delayInMsSteps()`
    to generate query intervals

  - Remove unused 'shallowFetch' from getAllTemplates() action
